refactor: utils パッケージを utils_core に改名
- src/utils を src/utils_core に改名
- tests/test_utils を tests/utils_core に改名
- 全てのインポートを utils → utils_core に更新
- pyproject.toml のパッケージ設定を更新

理由:
- プロジェクト内に複数の utils ディレクトリが存在（tests/database/unit/utils 等）
- pytest のパス解決で名前空間が競合していた
- utils_core という一意な名前により競合を完全に回避
