+++ pyproject.toml
@@ -86,7 +86,7 @@ requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [tool.hatch.build.targets.wheel]
-packages = ["src/analyze", "src/database", "src/market", "src/rss", "src/quant", "src/factor", "src/strategy", "src/automation", "src/news", "src/utils"]
+packages = ["src/analyze", "src/database", "src/market", "src/rss", "src/quant", "src/factor", "src/strategy", "src/automation", "src/news", "src/utils_core"]
 
 [tool.ruff]
 line-length = 88
